Updated error

+++ compile_images_to_train.py
@@ -6,13 +6,14 @@ import sys
 
 current_dir = os.path.dirname(os.path.realpath(__file__))
 training_classes = os.listdir('{}/datasets'.format(current_dir))
+invalid_files = ['.DS_Store', '.git', '.git_ignore']
 destination = 'train'
 
 for directory in training_classes:
 	class_dir = '{0}/datasets/{1}'.format(current_dir, directory)
 	for image in os.listdir(class_dir):
 		file_name = os.path.join(class_dir, image)
-		if (os.path.isfile(file_name)):
+		if (os.path.isfile(file_name) and file_name not in invalid_directories):
 			shutil.copy(file_name, destination)
 		else:
 			print('File doesn\'t exist: {}'.format(file_name))
